ENGDESK-14111 ENGDESK-14386 ENGDESK-14188 and more: Adjust DTMF timestamp
DTMF timestamp for RFC2833 packets is linear and monotonically increasing with regard to RTP timestamps.
This applies to when RTP timestamp is copied (SWITCH_RTP_FLAG_RAW_WRITE), and also when it is generated (rtp_rewrite_timestamps).
If a diff in time between last RTP packet and DTMF is more than milliseconds per packet
then DTMF's timestamp gets out of sync with a wall media clock (which is syncronised with RTP timestamp).

Even if this is a default behaviour of FreeSWITCH and it seems to be toletared most of the time
by RTP implementations (including FreeSWITCH itself), it introduces significant problems for some of them.
For example, Genesys call centre will drop DTMF due to misaligned timestamp.

This commit checks for gap in time between DTMF and last timestamp (of DTMF or RTP packet)
and adjusts DTMF timestamp for that. This has been tested with Genesys Cloud CX 2.